test: allow to capture test execution output
The tests can get their execution output appended to
a file by calling them with --logfile|-l

  test_foo.t -l /tmp/output.log

or by setting TEST_OUTPUT to that file.

Please note that for capturing the generated TAP output
there is already a dedicated variable TAPPER_OUTPUT; therefore
that output is not duplicated into the test output log.

Change-Id: I4d159acef63afbcea055c559aa1a562c5f6795f9