test: more test description on failing tests
When a test fails execution by early exit or
running into timeout we currently only report

 not ok - execution

This patch augments that line with the provided
arguments to tapper-wrapper and the test target:

  not ok - execution - run_test ex_stack_smashing

It covers several different use-cases of test
execution with BID tests, external tools, and other
scenarios. Instead of run_test the description could
also consist of external test execution tools and its
config file.

All arguments are stripped to their basename to focus
on the essentials instead of user-specific path names.

This helps to quickly identify which tests failed
without the need to read the full report, for instance
in a test summary page which lists only the fails
at the top.

Change-Id: I79138ddbd29b04a0466ff7d5ab5e8271dbf163c3