Fix PC file rule
Make sure that there is never a pattern rule matching a required PC
file from any lib other than the currently compiled one.

Without adding an explicit list of PC file targets it could happen that
the lib.mk for library B corrupts the PC file of library A because there
is a dependency to A's PC file and there is a pattern rule to create A's
PC file (with invalid contents).

Change-Id: Ic6e2a4689ecfcf95f47dfba1005e5056f4e7124c