test: tapper-wrapper writes logfile under prove
So far tapper-wrapper silenced STDOUT/STDERR to
/dev/null when run under prove ($HARNESS_ACTIVE).

For some errors that did not come from the captured
test STDOUT (like early errors in run_test) this
means the respective error messages got lost.

This patch appends that output to the already existing
logfile when requested with --logfile or $TEST_LOGFILE.

By that the overall behaviour of the quite complex file
handle redirection is untouched and no new output should
appear unexpectedly - there is just more content in the
logfile if requested.

Change-Id: I8b143d9a4625d0c85baf671fe16c9cee3c7d6574