Modules: Support fetching remote files via URLs
This adds support to fetch files via rsync/ssh and http/https
by supplying URLs in the modules.list. This is
very handy when building parts of the system elsewhere
or using common files like http-supplied ramdisks.

To only fetch files once per image build (or check once
whether they're up to date), we add functionality to
fetch the file or just get the local file name of
the cached file of a previous fetch operation.

Change-Id: I5e8f77e6eb36094f783de0d293ec81d24991e456