test: tapper-wrapper always expects Fiasco startup
Some tests work by printing a message in an infinite loop
and let tapper-wrapper wait for those expected lines.

When we run tests on physical machines that last infinite
loop might still be running there.

Now, when repeating the same test or a test expects the same
output as the last test, when we reconnect to such a system
we immediately see the expected message again although that
is not from our current test but still the old infinite loop.
The tapper-wrapper is happy with seeing this expected message
and ends immediately with a success.

This patch now always expects the first line to be from Fiasco
and therefore does not misinterpret old output.

When the output is expected repeatedly via TEST_EXPECTED_REPEAT
then it only rewinds the expected-output file and does not expect
the startup message again, so it works as before.

Change-Id: I3d9d39389fccda4715d14c87ce53edaee9588e61