genexportpack: allow network paths in grub entries
When generating grub config files for the exportpack the paths
are forced to be absolute paths. If they do not start with a
leading slash '/' then that '/' was prepended.

Now it also accepts leading paren '(' as absolute paths identifier,
like used in tftp network disk paths that look like '(nd)/some/path'.

Change-Id: Ia79abbdf788b90a97778e0ff0b2152661fc47dea