Add possibility to set the default architecture
When trying to create a build directory via B=path
and having the CROSS_COMPILE variable set, the
creation step does not work as the default
is being setup for x86 and thus the cross compiler
does not understand x86-specific options, and fails.

This adds support for considering an 'ARCH' environment
variable that allows to set the default architecture,
i.e. a
 export CROSS_COMPILE=arm-linux-
 make B=/tmp/dir ARCH=arm
will work then.

There's an alternative of using T=, e.g. with
T=arm-rv-v7a, however using the exact defconfig
names seems to be non-obvious for users.

Change-Id: I8751526bbef0054bc6ee6af20b943bcc5319e7b6