mk: fix distro detection in 'make report' cmd
The Linux distribution detection in the 'make report' command tests the
existence of certain files in /etc. However for Redhat and Slackware
only `[ -e /etc/*_version ]` without surrounding if/then is used. If the
file is not existent this will return 1 and 'make report' aborts.

This change fixes that by using if/then around the test. This fixes
'make report' on CentOS.

Change-Id: Ia5a8f0f247cd32411c9b8900c9a87eafeb8ec210