test: Introduce tags which can mark tests to be skipped by default
  - Adds TEST_TAGS to test.mk, where tags can be specified which are
    required in order to run the test.
  - Adds support for environment variable TEST_RUN_TAGS in run_test,
    which tags should be run.
  - Tags in TEST_TAGS that start with leading ! are checked to
    be NOT specified at run_test
  - Adds support for a cli options in the test script to specify
    which tags should be run.
  - Adds documentation for TEST_TAGS and the cli option.

This is helpful to skip tests that are slow, broken or should
not run by default for another reason, while still being able to
run the tests by specifying their tags in TEST_RUN_TAGS or using
the cli option --run-tags (or short -T) in the prove command
or when running the test script directly.

Test without any required tags are still run by default.

Tests with a "forbidden/negative" tag (starting with "!") are
run by default but not when the tag is provided in TEST_RUN_TAGS.

Example - Mark a test as slow in the Makefile:

  TEST_TAGS_test_target = slow,!gtest-shuffle

When running without the slow tag:

  $ prove -r test/t/
  test/t/path/to/test_target.t .. skipped: requires additional run tags: slow

When running with the slow tag:

  $ prove -r test/t/ :: --run-tags slow
  test/t/path/to/test_target.t .. ok

Running with forbidden tag:

  $ export TEST_GTEST_SHUFFLE=1
  $ prove -r test/t/ :: --run-tags slow,gtest-shuffle
  test/t/path/to/test_target.t .. skipped: not supported with run tags: gtest-shuffle

Change-Id: I86d2a7eb68a4b96da6d708f5eb484da398f9f66b