mk - Replaceable test starter
This change separates the configuration of tests from the
actual test execution which is generated in *.t test files.

The cleaner separation allows to provide a test starter script
as an alternative to the tapper-wrapper and run_test combination.

In order to stay backwards-compatible we introduce the script

  tool/bin/default-test-starter

which sets up the TMPDIR and runs tapper-wrapper and run_test
exactly as before.

In the same step we adjust test.mk to generate *.t files which
take a variable TEST_STARTER, which (in the *.t files) defaults
to above default-test-starter but can be replaced by setting the
variable beforehand.

When setting TEST_STARTER and running the .t file, it will set
the usual environment variables, like the search path or
TEST_TIMEOUT, which can also be utilized by alternative
test starters.

Change-Id: I7fbf1950adfde5ded0d546f9408e706b9467e442