use LC_ALL=C instead of LANG=C
When executing shell commands, use LC_ALL=C instead of LANG=C or
LANG=. The reason is that LANG is only used if the corresponding
LC_* environment variable is not set. For instance, the 'date'
command checks LC_TIME. And the general override for every LC_*
environment variable is LC_ALL. That is, the 'date' command uses
the content of LANG only if neither LC_TIME nor LC_ALL is set.

Change-Id: I8d581f56b143040eacdd9b7e8d99d8b2cc59e307