add -Wno-noexcept-type to hide annoying warnings
This gcc warning is about making "noexcept" part of the function type as
of C++17. Disable it because normally the whole tree is compiled with
the same gcc version. In the worst case an open reference to a function
could not be satisfied which can be fixed by recompiling everything.

Change-Id: I3cedf1fa81826aa315a394c1fc711af131d2c90c