Control: Allow to narrow 'requires' via expressions
This adds support to allow specifying expressions
to items in the 'requires' line of a Control file.
Examples:

  requires: acpica[arch(x86)] libc ..
  requires: libfoo[config(CONFIG_FEATURE_XYZ)]
  requires[arch(mips)]: libmips1 libmips2

Expressions are also possible:
  requires: acpica[arch(x86) or arch(amd64)]

The item will only be included if the expression evaluates to
true. Perl-like expressions are supported.

This mechanism is also available in alias defintions. There
is an additional keyword available to limit the dependency
to a build dependency, i.e. the library is not considered
when linking.

  foo-alias := libfoo[config(CONFIG_FOO);build-only]

Change-Id: I8de645bc840fded21acdaedadb984e0c9c799b9a
(cherry picked from commit a325269ba1326235d86c305c526561299af91685)