Make building tests optional via config option
This adds an additional TARGET type for test in the subdir role.
Targets for tests can be added with the TARGET_test variable.

These test targets are only built when the BID_BUILD_TESTS config option
is enabled. It is enabled by default or when the user explicitly calls
make in one of the subdirectories not otherwise descended into.

Adapt all the users of this accordingly.

Change-Id: Ia7590c83fa708993f02c9d35f9907142092fc4b3