Provide default bootstrap_cmdline_quoted value
When processing a modules.list (for example by qemu-x86-launch) that has
no bootstrap line in the entry and no default-bootstrap line in the file
the function get_module_entry of ModList.pm returns a hash with an unset
cmdline_quoted field for the bootstrap entry.

This may result in tools such as qemu-x86-launch reporting an error when
assembling the command line from these values.

This change provides a default quoted value.

Also the qemu-x86-launch tool passed the whole command line, including
the bootstrap name as argument to bootstrap. This changes splits the
command line removing the first space separated word which corresponds
to the bootstrap binary name.

Change-Id: I8dc06493887e52e5d7470f7321ae21e026ac05fe