Handle leading spaces in search path correctly
MODULE_SEARCH_PATH might contain leading spaces. If we apply a
simple split(/\s+/) to it we get an empty string as the first path.
This could lead to invalid search results, e.g. 'vmlinuz' could be
returned as '/vmlinuz' which  is a link to 'boot/vmlinuz*' on some
platforms.

Here we ignore empty strings returned by split.

Change-Id: I7d984803f94fb77c3a4626f3dda59a6988e3b7db