Provide meaningful error when gendep fails
We depend on preloading a library (libgendep) into the compiler to
generate dependencies. This does not work if the compiler is statically
linked. Provide a meaningful error message to alert users of the
problem.

Also check for the compilers to actually be present as part of the build
tools check.

Change-Id: I573fd7dcf5a87adf2fc559d00a9edaec0c0f46d6