{tool,mk} - Use /usr/bin/env in shebang for perl scripts
Many of the perl scripts in mk explicitly utilize the system's perl
interpreter by specifying /usr/bin/perl within the shebang line. Some
environments however have an alternative path to the perl binary or
prefer to use a different perl installation altogether such as a user
local one installed with perlbrew.

This change removes the restriction of using the system's perl
interpreter by starting perl via the env utility.

As many of these scripts use the -w or -W flag which can't be set when
using env due to certain limitations of the shebang, we remove these
flags and add "use warnings;" instead. Because use warnings only
activates warnings in the current scope this change additionally adds
use warnings to the perl modules in tool/lib. These are exclusively
loaded by perl scripts that had either the -w or the -W flag.

Change-Id: I2b1855b98bde3c04e52c8d02e8036a03465ef3c1