mk: Restructure Kconfig platform generation
The generation of the Kconfig file uses a large amount of shell code
used to generate the platform configuration options. This is currently
written directly in the Makefile which makes the code hard to read.
Move the shell code to its own script file.

Further, instead of post-processing the Kconfig file a second time for
generating platform definitions use standard source directives and
generate the sourced files containing the platform kconfig options.

Change-Id: Idba24691e8e4329fee5b37b00c874de94072923e