mk: Do not check build tools too early
Our list of build tools to be checked contains
variables that are set by ourselves and thus
checking the build tools cannot happen before
we actually created a build directory. For example,
CC will contain 'cc' initially (make's default)
which however is not required by our build (we use gcc
as the default) and thus leads to checking the
wrong programs.

Change-Id: I229cb4794543e2742182f33f37e83f3ee4f174e1