l4image: support editing ELF files
Move the image editing code into separate modules for each supported
file type. The details of the virtual address to file offset mapping
and the handling of the file specific data structures are encapsulated
there.

The ELF editing code assumes that the modules occupy exactly one
dedicated section. The module start address is verified to match the
start of a section. The end address is not known from the headers but
is assumed to be at the end of the section.

Because there is also data *after* the modules in ELF files the image
editing code will always work on a temporary file. After creating the
new output the temporary file is moved to the final location. If the
user gave no dedicated file name this will replace the input file.

Change-Id: Ifa88ee889026110c7e5717b1849cef9723d47578