mk: ony strip shared libraries
Static libraries and object files must not be stripped. Otherwise even
the non-stripped versions of binaries linked against those libraries
will miss some information that would otherwise be there. This changes
stripping to only happen when a file does have a .so suffix and the
BID_STRIP_PROGS option is enabled.

Change-Id: I3d1f5f0431e6f754ebb6b4d59411fbdb990cb907