tapper-wrapper - Fix confusing output
While using tapper-wrapper offline parts of the output are cut.
tapper-wrapper stops forwarding output to stderr as soon as TAP TEST
START is read, still forwarding whatever was in the buffer. After that
output to stderr is muted and tap lines are forwarded to TAPPER_FD.
When using tapper-wrapper locally this creates some confusion, because
a first part beginning with TAP TEST START is printed to stderr and
subsequently the actual relevant TAP result for prove is printed to
stdout. When running tapper-wrapper without prove stderr and stdout
end up in the same terminal. This can be rather confusing as it could
show duplicate ok or not ok lines without filtering out stdout, and
only a partial TAP TEST START/FINISH block when actually filtering out
stdout.

This change seeks to fix that, by continuing to forward output until
TAP TEST FINISH is found and gathering lines in between TAP TEST START
and TAP TEST FINISH in memory. After reaching TAP TEST FINISH, the
gathered lines are then printed to TAPPER_FD.

Change-Id: I4253cd75812ccc73abb23da5ba86baddfc32290b