Add fiasco config table and perl based full config combiner
To be able to detect fiasco's features in the test environment, we add
the contents of fiasco's globalconfig.out to a lua table in
test_env.lua.

As we don't want to to look into various tables later on, we do a
little bit of preparation work on the host side which enriches the
test_env.lua with a table of combined information from fiasco's table
and the HWCONFIG table. It is called t.FULLCONFIG.

The way this works is that run_test uses tool/lib/L4/TestEnvLua.pm to
collect information from TEST_HWCONFIG and TEST_FIASCOCONFIG if
present in form a perl hash. Run_test then extends the perl hash with
a few more information such as TEST_PROG. The whole hash is then
converted to lua code using the to_lua subroutine from
tool/lib/L4/TestEnvLua.pm.

For L4::TestEnvLua to be able to parse files the same way as
key-value-convert.pl we move the parsing mechanism to a perl module
called L4::KeyValueConfig, which contains a single subroutine called
"parse" taking one first boolean parameter, determining whether quotes
should be kept, followed by all the lines from the key value config to
parse. for key-value-convert.pl the first boolean parameter is 1, as
it expects quotes in the result, while for TestEnvLua the boolean
parameter is 0, because it needs to interpret the contents. It then
adds its own quotes around strings when converting the information to
a lua table.

Change-Id: Iee7d932621fd3ac311c6c7bcc8a782f0bbcc9ce5