Add support for kernel args specified by tests
This allows tests to set TEST_KERNEL_ARGS like any other TEST_*
variable. TEST_KERNEL_ARGS is then appended to the kernel_args within
run_test, so that the necessary kernel args for the test are provided
to fiasco even if the user uses the -d or -f parameter of
tapper-wrapper as well as the KERNEL_ARGS environment variable.

Change-Id: I99a29d113b4dcd61e9307ceeeb04fb45c1b7d3b8