tapper-wrapper - Add support for a hard timeout
The current timeout tapper-wrapper supports only works for cases where
the test does not output anything, and resets as soon as something is
read. This works well, however it does not prevent tests from
endlessly flooding the log without actually progressing. To counter
these cases we add an additional timeout, we call hard timeout, that
limits the time the test can run. We expect to set this to an
arbitrarely high value, like 30 minutes or 1 hour, to prevent the
extreme cases which unexpectedly flood the log for ~10 hours.

Change-Id: I7dcd95d13babb01592d4824293102df387db9fab