tapper-wrapper - Prevent printing TAP twice
When running a t-file without prove tapper-wrapper now prints TAP
twice, which is unnecessary if the running test does not use the
bundle mode. In case the test uses the bundle mode, it appears like a
nice to have to see the combined tap output again at the end.

So to prevent that we detect the usage of prove via the HARNESS_ACTIVE
variable and skip printing the TAP lines again if prove is not
detected and the test did not use the bundle mode.

Change-Id: I7892b240b3e13fb81c6fe7e1512d220e2583811d