tool/l4image: Handle wrapped bootstrap elf images
On amd64 the bootstrap image is a 32bit image that contain another
bootstrap image which is 64-bit and launched after the mode-switch.

The l4image tool needs to deal with the special case because the inner
file is the one to actually be analyzed / modified.

We solve this by extracting the inner elf binary if one is detected,
manipulating that file and then replacing the section where that file
resided after we are done.

Change-Id: I1119784cdaac5220fef08f6d51d9aa2d4c5520c1
Requires-Change-Id: I0c7bedbfe2db3d1a22abcd741860bc1fba0101c2