bid-bender.spec: support --wrap=symbol
Using --wrap=foo the linker wraps all calls to foo to __wrap_foo(). That
function in turn may call __real_foo() to call the original function.

Using this mechanism the implementation of dedicated functions can be
changed. It is also possible to execute code before / after the original
function is executed.

Change-Id: I488e82712b2a3a2212bc44c657be0a8273dffdaa