tapper-wrapper - Allow disabling line-wise timeout
Some tests are expected to be silent for a longer period of time.
Instead of setting the TEST_TIMEOUT ridiculously high, it is a cleaner
solution to allow disabling the timeout between lines. This change
introduces the necessary steps to call can_read with undef if the
environment variable TEST_TIMEOUT is set to 0.

Change-Id: I680c85a5283b2bc94224b79bbb6614bba65f6e32