run_test - Extend test tag support
We realized that the current test tag mechanism does not fit our needs
completely. Until now test tags could either be specified or
unspecified which means that tagging tests disables them by default.
While there is an ! operator for tags which skips test if a certain
tag is specified, it would lead to confusing double negatives.

To solve this problem we first extend the states a tag can have. While
previously a tag could be specified or unspecified, now a tag can be
set to 1, set to 0 or be unspecified. With this the TEST_RUN_TAGS
variable now can contain tag=1, tag=0 or not mention the tag at
all (unspecified). For backward compatibility reasons tags specified
in TEST_RUN_TAGS without either =0 or =1 are automatically assumed to
be set to 1.

As a next step we add 2 more notations for tests to describe
conditions on run tags: the + and the - notation. The following table
shows the 4 notations with their conditions how the tag was or was not
specified within TEST_RUN_TAGS.

TEST_TAGS   | TEST_RUN_TAGS contains
tag         | 'tag=1' or 'tag'
!tag        | unspecified or 'tag=0'
+tag        | unspecified or 'tag=1' or 'tag'
-tag        | 'tag=0'

To clear up their use cases let's apply them to a few examples for
tags.

The first tag there ever was is long-running, which we want to be
explicitly specified if a test tagged as such ought to be run. To
achieve this we specify the tag in the Makefile without any prefix.

Another tag is hardware, which is used to mark tests that must not be
run on hardware by using the ! notation. Testruns on hardware set
TEST_RUN_TAGS to hardware, so that tests specifying !hardware are not
run. Testruns on qemu do not specify hardware and tests tagged with
!hardware are therefore executed.

A new use case we introduce with this change are tags such as impl-def
which shall be used with the + notation. These tags mark tests with a
certain category without disabling their execution by default, but
gaining the ability to explicitly skip them by specifying impl-def=0
within the TEST_RUN_TAGS variable.

For completeness sake we also add the - notation for tests that ought
only to be run if a tag is explicitly set to 0, so tests which only
ought to be run if a certain condition designated by a tag is not met,
while skipping the test as well if unspecified.

Fortunately the new set of notations is backwards compatible,
therefore we do not have to alter any tags already in use.

Change-Id: Ia65f21ef182150f40c7cbc0574ed074fea4d88a2