tapper-wrapper - Handle TEST_TIMEOUT if empty but defined
In case a TEST_TIMEOUT is empty but defined it's value stayed
a string instead of applying the default integer 10, which causes
problems in readline_timeout. To fix this we test not just whether
TEST_TIMEOUT is defined but also consider if it's a number or not when
applying the default 10 seconds.

Change-Id: I331ffeed7046e6012c32331c7a4075329c180fca