mk: remove potentially unavailable libraries/objects
Bare metal toolchains do not have the libgcc_eh.a library and miss some
of the crt*.o object files. Make the former optional during linking.
Remove the latter from the Makefile because there is no user in the
system anymore which requires the absolute path to the crt*.o files.

Change-Id: Id898d084a94ab8d2dbfb438e9fdf493212fe9cc9