conf: add lua lib to parse hwconfig for tests
The lua library is included in the modules.list for use in a ned.cfg.
It can be required in a config script under the name: "rom/lib_test".

Failures due to wrong usage of the library, e.g. passing an integer
where a string is expected lead to a TAP formatted 'not ok' line
containing the error message.

Failures due to a missing HWconfig, missing property, or a property not
having the expected value print a TAP formatted 'ok-skip' message.

Change-Id: I7b02cc55ba2fbe88cd99d5986bd860e77d9ffa18