Makefile: Do not update RAM_BASE after an explicit switch_ram_base
Until now RAM_BASE is set to PLATFORM_RAM_BASE everytime an image is
generated, which renders the switch_ram_base goal unusable. To fix
that we check via RAM_BASE_SWITCH_OK if the RAM_BASE was set
explicitly and skip the update in that case.

As we want the explicit RAM_BASE switch to be ignored whenever we
change the PLATFORM_TYPE, we record the PLATFORM_TYPE within the
Makeconf.ram_base file, so that we can detect if the explicit RAM_BASE
was set for another platform than we're using for the image.

Change-Id: I87dd2a05e60599701055e46d6d8dc4358a40b712