run_test: Use system instead of piping from subprocess to stdout
The section for running an EXTERNAL_TEST_STARTER simply starts a
subprocess and pipes its stdout to run_test's stdout. This can simply
be replaced by a call to the system function. For cases where system
does not find the external test starter, we print a special error
message, otherwise we just exit with the same exit code as the
external test starter.

Change-Id: Ieb04adac2f967466266b471166d45a0841d0c0c1