Add generator for kernel unit test starter files
Pre: The kernel was compiled with the unit test configuration enabled
and there are test_* files present in the MODULE_SEARCH_PATH.

This introduces a makefile target 'kunit-test' which prepares
the environment before it invokes the new script 'gen_kunit_test'.

gen_kutest searchs the for the kernel binary and generates for all
test_* files a .t-file containing the environment to actually
start a unit test.
The resulting .t-files are places into the build directory under
test/kunit/.

If you specify KUNIT_VERBOSE=1, all generated file names will be
printed.

Change-Id: I6e6211305517b1ead2f34596f258efdcae3d0cfb