Recompile C/C++ only if necessary
This changes how C and C++ is considered for
recompilation by tracking information on each
object file to build and only consider recompilation
if any of those prerequisites change. This is the
case of the command to compile or any of the direct
dependencies change. Recompilation is not required
if unrelated files, such as the build system or
unrelated configuration change.

Change-Id: I62ae36a1dffedc62457603ab731219070148db23