l4image: Add script-based standalone variant
Add a l4image generation variant that
just creates a script. It is much smaller
but requires to have libpar-perl installed.
Also move our Digest::CRC copy to our namespace such
that it does not clash with the real Digest
namespace.
For generating the binary variant, it is good
to do this on a system with an ancient glibc
such that it works on many systems out there.

Change-Id: Ida43aacc8dfc3a31a91b3d94e0fde13f25ec7a09