tapper-wrapper - Add limit for number of lines read
In some cases new tests can end up in endless loops of fast output
leading to no result. Instead of waiting for the timeout to happen and
ending up with +1GB log files, we limit the lines read from tests and
terminate the test if that limit is reached. As default we limit the
maximum lines to a hundred thousand. This can be overwritten either by
setting the environment variable MAX_LINES_READ or providing the flag
--max-lines-read n to the .t file.

Change-Id: I211afa5d554d1a5e092d74188a026b64a0a9057e