bid/kconfig: Add experimental option
We sometimes have features that are not fully ready for everyday use but
that should still already be included for broader testing or to
expedited development. Add an option to guard these features with an
appropriate warning in the description.

If you enable it and something breaks during compile or run-time, you
are on your own.

Change-Id: Iec49f263eb2a18abaa5f6d9ccb8f34bc919b4149