mk: add support for execute-in-place
Add the notion of a flash image that can be created by the flashimage
target. Images that are marked as xip in modules list, e.g.

  module[xip] init

are placed at the right spot in the flash image to be executable
directly. This requires that the executable is linked to the flash
executable address (Makeconf.local):

  XIP_init              := y
  DEFAULT_RELOC_init    := 0x00f40000
  DATA_RELOC_init       := 0x00040000

If XIP_* is set the DEFAULT_RELOC_ offset is into the
PLATFORM_FLASH_BASE address instead of PLATFORM_RAM_BASE. The
DATA_RELOC_ offset is to place the data/bss section at the desired
place counted from PLATFORM_RAM_BASE.

Change-Id: I1735a8a3ee6dbd1bf24706a4d6c39cc4528dbf30