tapper-wrapper - Add feature to prefix output with timestamps
To get information about the timeline of a test execution we add a run
time in seconds in front of every line of the captured output that is
written to the testout.log or the terminal. The checks for expected
output and the TAP block handling loop utilize the unaltered output
and thus work as before. The feature can either be enabled by adding
the --timestamps cli option or by setting the TEST_TIMESTAMPS
environment variable to 1.

Change-Id: Ib9ba4eb44ad432970bff4932b5edebf3fffb0f03