tap(per)-wrapper: Move to a plugin architecture
This change makes tapper-wrapper more flexible for future additions to
different test and analysis infrastructure. Output can be post-processed
by plugins that may generate additional TAP lines.

Also, the name tapper-wrapper was not the best choice, since it wraps
the TAP output (and not tapper). Thus rename the tool while we add it
but keep a symlink for backwards compatibility).

Change-Id: I276fed9b19868091ea54cdf071dc487a4dd83375