arm: Handle -march option for recent gcc's
With gcc-11, at least with Debian's cross tool chain, one cannot specify
-march=armv7-a anymore for a target that shall contain FPU code as the
underlying call to cc1 mandates a hard-float API and thus refuses to
compile.  Thus we use -march=armv7-a+fp to say we also want floating
point. However, earlier gcc do not yet support the '+fp' so we have to
figure it out in the config phase.

Change-Id: Ib33b9a0903baa754c673e85e9e6cdb5b49baa4ee