tool/bin/run_test - Retrieve selected platform through Makeconf
The approach to export the PLATFORM_TYPE into the l4defs during build
for run_test to read does cover all use cases. We decided that a
better approach is to retrieve it via mk/Makeconf directly. By using
mk/Makeconf directly we circumvent any side effects that might cause
issue if run_test were to be run in parallel.

Change-Id: I33a7aa60e2d83a30dfe60a75450cd9bbf8bb8c45