arm64: Add check and use -mno-outline-atomics
Recent libgcc use a glibc/Linux-provided flag to decide whether to use
LSE or LLSC atomics. This check is implemented at runtime with a
function in libgcc. Code that does not link against libgcc fails to
link. Further, we do not give the flag, so only LLSC will be taken
anyway. Use -mno-outline-atomics to avoid gcc generating the choice and
thus the link failure when not linking with libgcc.

Change-Id: I4bada79b420f25dc08b27bdf831a4ca837b7d521