BID: Fix generation of linker name
Happened that calling LD with an invalid emulation name yields to an
error message that we cannot parse with the given expression. The
emulation name is only determined after that. Now it is hard to use even
if reordering the sequence. Thus just use the same trick as below and
call the linker without any argument to avoid this in the first place.

Change-Id: I571794c7d70cd3a8ae56724dbb0ade4be6f223f0