mk: unify GCC linker flag conversion
The conversion from ld-style to gcc style linker flags must preserve
the order of flags that must not be converted compared to flags that
must be converted. Adapt the function to do this.

Also unify the same block from lib.mk such that the listing of flags to
be converted must not be maintained in two places and move it to
Makeconf to make the function more easily accessible to other users.

Change-Id: Iff13238ee8b7873396aa1fa98983b1807b1ca4a2