Tests prefer testfile over target in fail messages
So far the tap-wrapper by default uses the target binary
name as test description when reporting an execution
fail.

While this is more precise when a single test.t
runs with multiple different binaries, it makes the
test more difficult to find when reading error reports
because the test file.t name is what the developer
usually needs when digging into problems, especially
when they failed in more dramatic ways like a timeout
where only few information is available.

Now we use the actual testfile name where possible,
in bid-tests and kunit-tests.

Change-Id: I7dad524037d53d7f8ec405111929401c459480b9