gen_kunit_test: Allow more flexible config
Currently only expected files are possible, however in order to be able
to more flexibly configure test output processing plugins it must be
possible to specify more configuration options.

Thus move the config for a test to the config_<testname> file which will
then be inlined into a generated .t file. Also move that inlining to the
end so that the config may override default values.

Change-Id: I49f7e7d0fd1d70398c0d4ae5965d30df8739edea