tapper-wrapper - Fix timestamps
Putting a timestamp after every newline we encounter leads to a
timestamp not showing when the line started but rather when the
previous line ended. This is especially confusing when there's a
larger amount of time between the previous line and the current one.
To prevent this confusion we adjust the timestamp feature to add a
timestamp before the first character after a newline.

There's two cases we have to handle while examining chunks. One case
is simple: a newline and the following character are in the chunk,
then we can just insert a timestamp. The other case is that a newline
was at the end of one chunk and the following character is a the
beginning of the next chunk. To handle that we set a variable if the
current chunk ends in a newline, and insert a timestamp in the
beginning of the next chunk if the variable was set.

Change-Id: I908a37d2149bfa15ecced22a44b4433bf7803012