tap-wrapper: Allow specifying additional plugins
Plugins that are specified either through the TEST_TAP_PLUGINS
environment variable or as argument to the .t file are now loaded in
addition to the normal plugins loaded for the file.

The default plugins are set by the .t file unless the Makefile
explicitly specifies plugins that shall be used.

Change-Id: Ib8e65f31beeea16c6312b45f0154870b71823858