tap-wrapper: Deal with prefixes in tag plugins
Enable processing of tagged lines even if they are prefixed by some
arbitrary string. If we deal with a tagged block then remember the
string and make sure that all lines in the block are only counted if
they also have that prefix.

This mirrors the behavior for classic TAP output lines.

Change-Id: Id7d62c5c14d125cfa7b248d61705f42e8dc92feb