tool/bin/tap-wrapper - Fail early if no plugins were loaded
In the case that no plugins were loaded tap-wrapper currently simply
skips the test without message. This is dangerous since we just
removed the old default for plugins and cannot be sure that all tests
now explicitly provide the list of plugins they need. To find all of
them we simply fail if no plugins were loaded.

Change-Id: I625b5a9826575ea39c6e4e63daa51c042142e846