mk: support statically allocated stack no-MMU systems
It is usually a bad idea to dynamically allocate RAM on no-MMU systems.
Instead, the linker script of the executable can reserve some space in
the BSS section. The stack size can be adjusted through the
DEFAULT_STACK_SIZE variable.

This part only passes the information to the linker by a special
symbol. It's up to the platform to adapt the linker script to make use
of this. Additionally, the loader also needs to make use of the
reserved stack region.

Change-Id: If7dc218004d41651fc5d986720536a6d4aa3d7fc