tap-wrapper: Plugin for invoking external tools
This change adds a tag-based plugin that can be used to invoke an
external tool. The tool gets passed a directory to snippet information
that belong to the specified tag.

Every tagged block and line are written to a file in that directory,
without their tag information. The files are numbered in sequence as
found in the test output and contain the tag as well as additional info
specified for the block or tag.

Change-Id: Icf67aecb9872dfc6efc6a234ee955c689144700b