ARM/ARM64/x86/AMD64: detect no-FPU compiler flags
These flags shall prevent the compiler from generating FPU instructions
or instructions for similar processor units (SSE, AVX, etc). Be aware of
library code when assuming that these parameters are used. It is usually
not possible to specify these flags for code depending on libstdc++
features.

On ARM64, unconditionally use '-mgeneral-regs-only' which was introduced
with the initial support of that architecture in gcc-4.8.

On ARM, detect '-mgeneral-regs-only' which is supported as of gcc-9.

On x86/AMD64, detect various parameters to disable MMX/SSE/AVX code.
Some parameters have already been recognized on x86.

Change-Id: I063ad1b77258703980da92428be47df1a03ed922