mk: add armv7plus and armv8plus definitions
The definition of the arm architecture revision (v6, v7, v8, ...) is
orthogonal to the execution mode (AArch32, AArch64). The only
dependency is that AArch64 is defined starting with Armv8.

Because both Kconfig.arm.inc and Kconfig.arm64.inc are sourced, the
common definitions must be moved to a separate file.

Change-Id: I6838bfc0c3dd17f85f7a7fad46187eb01898d7b9