concepts: document when granting is not performed
If the resulting attributes of a flexpage as they would appear in the
receive window either don't have the L4_CAP_FPAGE_R bit (object caps)
or none of the L4_FPAGE_RWX bits (memory, IO ports), then the grant
operation is not performed: The flexpage is not removed from the sender
and is not added to the receiver.

Link to the L4_MAP_ITEM_GRANT flag where this is documented.

See ticket #CD-689.

Change-Id: Id9ba3969f687a9ba528a616c98997f9b2045f574