don't try to use '--hash-style=sysv' for LDFLAGS any longer
The detection of the presence of the '--hash-style' option did not work
for ages because when we evaluate $(LD), the LD_EMULATION option was not
yet known which resulted in an invalid call of the linker because

  LD = $(CROSS_COMPILE)ld -m $(LD_EMULATION) --oformat $(OFORMAT)

So actually '--oformat' was taken as linker emulation and the command
failed always.

Setting the hash style to some legacy format is not necessary anyway
because in the meantime, all tools should be able to understand the
current format.

Change-Id: I50b98fb7295d344030d05ab2f95bc347092cbe83