tap-wrapper - Do not steal stdin unless necessary
For automated tests we introduced a mode where tapper wrapper dumps
the JDB system report on a timeout. For this we needed a dedicated
pipe to the child processes stdin. This however collides with local
usage because a developer cannot access the child processes stdin
anymore without supplying the -W flag. This change fixes that by only
creating a dedicated pipe for the child's stdin if dumping the JDB
system report is requested.

Change-Id: Ie0b9523421bca7150b35af63b53124a15255c722